+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/sampler/_aws_sampling_result.py
@@ -1,0 +1,37 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+from types import MappingProxyType
+from typing import Optional
+
+from amazon.opentelemetry.distro._aws_attribute_keys import AWS_XRAY_ADAPTIVE_SAMPLING_CONFIGURED_ATTRIBUTE_KEY
+from opentelemetry.sdk.trace.sampling import Decision, SamplingResult
+from opentelemetry.trace import TraceState
+from opentelemetry.util.types import Attributes
+
+
+class _AwsSamplingResult(SamplingResult):
+    AWS_XRAY_SAMPLING_RULE_TRACE_STATE_KEY = "xrsr"
+
+    def __init__(
+        self,
+        decision: Decision,
+        attributes: "Attributes" = {},
+        trace_state: Optional["TraceState"] = None,
+        sampling_rule_name: Optional[str] = None,
+        has_adaptive_sampling_config: bool = False,
+    ):
+        super().__init__(decision, attributes, trace_state)
+
+        self.attributes = MappingProxyType(
+            {
+                **self.attributes,
+                AWS_XRAY_ADAPTIVE_SAMPLING_CONFIGURED_ATTRIBUTE_KEY: has_adaptive_sampling_config,
+            }
+        )
+
+        if self.trace_state is None:
+            self.trace_state = TraceState()
+        if self.trace_state.get(self.AWS_XRAY_SAMPLING_RULE_TRACE_STATE_KEY) is None:
+            self.trace_state = self.trace_state.add(self.AWS_XRAY_SAMPLING_RULE_TRACE_STATE_KEY, sampling_rule_name)
+
+        self._sampling_rule_name = sampling_rule_name